+++ lib/src/metta/runner/stdlib/core.rs
