+++ lib/src/metta/runner/stdlib/string.rs
